+++ lib/IRequestResponse.d.ts
@@ -1,6 +1,6 @@
 import { Request } from './Request';
 import { Response } from './Response';
-export interface IRequestResponse {
-    request: Request;
-    response: Response;
+export interface IRequestResponse<TRequest = any, TResponse = any> {
+    request: Request<TRequest>;
+    response: Response<TResponse>;
 }