+++ lib/ResponseData.d.ts
@@ -1,8 +1,8 @@
 import { StatusCode } from './StatusCode';
-export declare class ResponseData {
+export declare class ResponseData<TData = any> {
     private status;
     private data;
-    constructor(status: StatusCode, data?: any);
+    constructor(status: StatusCode, data?: TData);
     getStatus(): StatusCode;
-    getData(): any;
+    getData(): TData;
 }