+++ src/IRequestResponse.ts
@@ -16,7 +16,7 @@
 import {Request} from './Request';
 import {Response} from './Response';
 
-export interface IRequestResponse {
-    request: Request;
-    response: Response;
+export interface IRequestResponse<TRequest = any, TResponse = any> {
+    request: Request<TRequest>;
+    response: Response<TResponse>;
 }